Move pairwise to util.py
